… src/Columns/ColumnsCommon.cpp