### Changelog category (leave one):
- Performance Improvement


### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Improve performance of ColumnArray::filter for big int and decimal


#### Optimize ColumnArray::filter
##### Int128
Q: SELECT arr FROM (SELECT cast(range(number % 10) as Array(Int128)) AS arr FROM (SELECT * FROM system.numbers LIMIT 100000000) WHERE length(arr) <= 5) format Null; 
Before: 0 rows in set. Elapsed: 3.030 sec. Processed 100.01 million rows, 800.08 MB (33.01 million rows/s., 264.05 MB/s.)
After: 0 rows in set. Elapsed: 2.396 sec. Processed 100.01 million rows, 800.08 MB (41.74 million rows/s., 333.89 MB/s.)
Speedup: 1.26x

##### Int256
Q: SELECT arr FROM (SELECT cast(range(number % 10) as Array(Int256)) AS arr FROM (SELECT * FROM system.numbers LIMIT 100000000) WHERE length(arr) <= 5) format Null; 
Before: 0 rows in set. Elapsed: 4.522 sec. Processed 100.01 million rows, 800.08 MB (22.12 million rows/s., 176.94 MB/s.)
After: 0 rows in set. Elapsed: 3.967 sec. Processed 100.01 million rows, 800.08 MB (25.21 million rows/s., 201.67 MB/s.)
Speedup: 1.14x 

##### Decimal64
Q: SELECT arr FROM (SELECT cast(range(number % 10) as Array(Decimal64(0))) AS arr FROM (SELECT * FROM system.numbers LIMIT 100000000) WHERE length(arr) <= 5) format Null; 
Before: 0 rows in set. Elapsed: 5.048 sec. Processed 100.01 million rows, 800.08 MB (19.81 million rows/s., 158.51 MB/s.) 
After: 0 rows in set. Elapsed: 4.494 sec. Processed 100.01 million rows, 800.08 MB (22.25 million rows/s., 178.01 MB/s.)
Speedup: 1.12x


##### Decimal128 
Q: SELECT arr FROM (SELECT cast(range(number % 10) as Array(Decimal128(0))) AS arr FROM (SELECT * FROM system.numbers LIMIT 100000000) WHERE length(arr) <= 5) format Null;  
Before: 0 rows in set. Elapsed: 11.869 sec. Processed 100.01 million rows, 800.08 MB (8.43 million rows/s., 67.41 MB/s.)
After: 0 rows in set. Elapsed: 11.079 sec. Processed 100.01 million rows, 800.08 MB (9.03 million rows/s., 72.22 MB/s.)
Speedup: 1.07x 


##### Decimal256
Q: SELECT arr FROM (SELECT cast(range(number % 10) as Array(Decimal256(0))) AS arr FROM (SELECT * FROM system.numbers LIMIT 100000000) WHERE length(arr) <= 5) format Null; 
Before: 0 rows in set. Elapsed: 34.749 sec. Processed 100.01 million rows, 800.08 MB (2.88 million rows/s., 23.02 MB/s.)
After: 0 rows in set. Elapsed: 34.321 sec. Processed 100.01 million rows, 800.08 MB (2.91 million rows/s., 23.31 MB/s.)
Speedup: no benefits, probably because `cast(range(number % 10) as Array(Decimal256(0)))` took too much time, which weaken the benifits of this PR. Maybe this is another performance issue we need to solve.