@taiyang-li 2023-02-02
#### Optimize ColumnArray::replicate
##### Int128
Q: with cast([1,2,3,4] as Array(Int128)) as elem select arrayWithConstant(rand() % 10 + 5, materialize(elem)) from numbers(1000000) format Null
Before: 0 rows in set. Elapsed: 0.696 sec. Processed 1.05 million rows, 8.37 MB (1.50 million rows/s., 12.03 MB/s.)
After: 0 rows in set. Elapsed: 0.463 sec. Processed 1.05 million rows, 8.37 MB (2.26 million rows/s., 18.09 MB/s.)
Speedup:  1.50x 

##### Int256
Q: with cast([1,2,3,4] as Array(Int256)) as elem select arrayWithConstant(rand() % 10 + 5, materialize(elem)) from numbers(1000000) format Null
Before: 0 rows in set. Elapsed: 1.088 sec. Processed 1.05 million rows, 8.37 MB (961.69 thousand rows/s., 7.69 MB/s.)
After: 0 rows in set. Elapsed: 0.853 sec. Processed 1.05 million rows, 8.37 MB (1.23 million rows/s., 9.82 MB/s.)
Speedup: 1.28x

##### Decimal64
Q: with cast([1,2,3,4] as Array(Decimal64(0))) as elem select arrayWithConstant(rand() % 10 + 5, materialize(elem)) from numbers(1000000) format Null
Before: 0 rows in set. Elapsed: 0.290 sec. Processed 1.05 million rows, 8.37 MB (3.61 million rows/s., 28.92 MB/s.)
After:  0 rows in set. Elapsed: 0.111 sec. Processed 1.05 million rows, 8.37 MB (9.39 million rows/s., 75.12 MB/s.)
Speedup: 0.26

###### Decimal128
Q: with cast([1,2,3,4] as Array(Decimal128(0))) as elem select arrayWithConstant(rand() % 10 + 5, materialize(elem)) from numbers(1000000) format Null
Before: 0 rows in set. Elapsed: 0.653 sec. Processed 1.05 million rows, 8.37 MB (1.60 million rows/s., 12.83 MB/s.)
After: 0 rows in set. Elapsed: 0.473 sec. Processed 1.05 million rows, 8.37 MB (2.21 million rows/s., 17.69 MB/s.)
Speedup: 1.38x

###### Decimal256 
Q: with cast([1,2,3,4] as Array(Decimal256(0))) as elem select arrayWithConstant(rand() % 10 + 5, materialize(elem)) from numbers(1000000) format Null
Before: 0 rows in set. Elapsed: 1.025 sec. Processed 1.05 million rows, 8.37 MB (1.02 million rows/s., 8.17 MB/s.)
After: 0 rows in set. Elapsed: 0.982 sec. Processed 1.05 million rows, 8.37 MB (1.07 million rows/s., 8.53 MB/s.)
Speedup: 1.04x